+++ oalg-top/src/OAlg/Homology/ChainComplex.hs
@@ -26,6 +26,7 @@ module OAlg.Homology.ChainComplex
   ) where
 
 import Data.Type.Equality
+import Data.Foldable (toList)
 import Data.List as L (zip)
 import Data.Maybe
 
@@ -48,7 +49,6 @@ import OAlg.Entity.Matrix
 
 import OAlg.AbelianGroup.Definition
 
-import OAlg.Homology.Simplex
 import OAlg.Homology.Complex
 
 --------------------------------------------------------------------------------
@@ -92,30 +92,31 @@ instance Distributive a => Validable (ChainComplex t n a) where
 --------------------------------------------------------------------------------
 -- chainComplexZ -
 
-chainComplexZ :: Ord v => Complex n v -> ChainComplex From n (Matrix Z)
+chainComplexZ :: Simplical s x => Complex s n x -> ChainComplex From n (Matrix Z)
 chainComplexZ c = case chain c of
   DiagramChainFrom n ds -> ChainComplex (DiagramChainFrom dZero (zero (dZero :> n) :| ds))
   where
 
     dZero = one ()
 
-    chain :: Ord v => Complex n v -> Diagram (Chain From) (n+2) (n+1) (Matrix Z)
-    chain (Vertices vs) = DiagramChainFrom n (zero (n :> dZero):|Nil) where n = dim () ^ lengthN vs
+    chain :: Simplical s x => Complex s n x -> Diagram (Chain From) (n+2) (n+1) (Matrix Z)
+    chain (Vertices s) = DiagramChainFrom n (zero (n :> dZero):|Nil) where n = dim () ^ lengthN s
     chain (Complex ss c) = case chain c of
       DiagramChainFrom n ds -> DiagramChainFrom m (d:|ds) where
         m = dim () ^ lengthN ss
         d = Matrix n m (rcets $ rc (listN ss) (cplIndex c))
 
-        rc :: (N ~ i, N ~ j)
-          => [(Simplex (n+1) v,j)] -> (Simplex n v -> Maybe i) -> Row j (Col i Z)
+        rc :: (Simplical s x, N ~ i, N ~ j)
+          => [(s (n+1) x,j)] -> (s n x -> Maybe i) -> Row j (Col i Z)
         rc ss f = Row $ PSequence $ amap1 (colj f) ss
 
-        colj :: Ord i => (Simplex n v -> Maybe i) -> (Simplex (n+1) v,j) -> (Col i Z,j)
-        colj f (s,j) = (col f (faces' s),j)
+        colj :: (Simplical s x, Ord i)
+          => (s n x -> Maybe i) -> (s (n+1) x,j) -> (Col i Z,j)
+        colj f (s,j) = (col f (toList $ faces s),j)
 
-        col :: Ord i => (Simplex n v -> Maybe i) -> [Face (n+1) v] -> Col i Z
+        col :: Ord i => (s n x -> Maybe i) -> [Face s (n+1) x] -> Col i Z
         col mf fs = colFilter (not.isZero) $ Col $ psequence (+) (alt `zip` amap1 (f mf) fs) where
-          f :: (Simplex n v -> Maybe i) -> Face (n+1) v -> i
+          f :: (s n v -> Maybe i) -> Face s (n+1) v -> i
           f m (Face s) = case m s of
             Just i -> i
             _      -> error "inconsistent complex"
@@ -127,9 +128,9 @@ chainComplexZ c = case chain c of
 --------------------------------------------------------------------------------
 -- chainComplex -
 
-chainComplex' :: (Hom Dst h, Ord v) => h (Matrix Z) a -> Complex n v -> ChainComplex From n a
+chainComplex' :: (Hom Dst h, Simplical s x) => h (Matrix Z) a -> Complex s n x -> ChainComplex From n a
 chainComplex' h c = ccplMap h (chainComplexZ c)
 
-chainComplex :: Ord v => Complex n v -> ChainComplex From n AbHom
+chainComplex :: Simplical s x => Complex s n x -> ChainComplex From n AbHom
 chainComplex = chainComplex' FreeAbHom